refactor: remove never-ending loop in ScanHandler
In ScanHandler::_perform_scanning_logic(), we were calling the method
itself if we initially could not send our request due to not having
control, but then succeed at getting it. It is unlikely to happen, but
its *possible* that we could consistently get and lose control between
these recursive calls, creating an endless loop.

I admit, this is quite unlikely. Perhaps the better reason for
refactoring this is simply that it is less confusing. There is no real
reason to introduce recursion here when its not necessary.