fix: get all UTs running at once for controller
When running the controller tests simultaneously with gxsm, we were
getting random errors. After debugging, we realized the issue is due to
ZMQ and its handle of the context. Particularly, even though the
client *should* be deleted between tests, it does not appear to be
guaranteed that it is.

For now, we are explicitly deleting the client between tests. If this is
insufficient, we can call client._close_client() instead.

Longer term, we may want more proper handling of explicit component
closure. This is less of an issue live, because the component monitor
handles closure properly.