fix: get test_controller running reliably
Previously, we called del client at the end of each test. It turns out
that this was not enough! For now, what we are doing is:
- call _close_client() to explicitly sever the connection;
- call del client to explicitly delete the object;
- sleep for a little bit (1 second).

This appears to work. It's hack-ey, but should be OK for now, since this
issue should not show up when used live (see comment description).