feat: delete pre-existing handlers before logging
This is a bit of a weird feature. If you are running logging multiple
times in the same python interpreter, it is possible the previous run's
handlers have not been destroyed. This is why you should generally
remove handlers you have added at the end of your run.

I *did not* do this, which is not great. However, this problem had not
really arisen before, since our system usually only runs 1x per process.
When testing with gxsm, we ran into an unusual situation:
a) We could be running a component multiple times without resetting the
interpreter (because we are running inside an embedded python
interpreter);
b) Gxsm adds some stdout/stderr redirection for each run which gets
destroyed between runs.
c) However, since we were not deleting our added handlers, the deleted
redirection still exists in the logger! This was causing weird printout
errors to be shown.

The fix is to clear our (prior) handlers whenever we set up for a new log.