fix: update dependencies
- We were using gxsmread as a dependency in the 'gxsm' group, but it was
not optional! We have added it as an optional dependency now, tied to
extras 'gxsm'. Note that this is currently referencing a local
path, *requiring* that you have gxsmread cloned into ../gxsmread. This
will be changed eventually, once it is packaged into PyPI.
- Replacing unused dependencies with the true dependencies we need. So,
for example, we do not use (nor require) sidpy, but *do* require scipy
and matplotlib.
- Updating lock file accordingly.