fix: Update multiprocessing usage for all OSes
Creating a new process functions differently between OSes! Unix uses
'fork', which directly copies the current state of all variables/etc.
Windows and Mac OSX currently use 'spawn', which *does not* do this. Put
differently, the 'startup' of global variables/etc is re-done when a
process is spawned. Since Windows *cannot* 'fork', we cannot use the
existing approach on all OSes.

To be consistent across methods, I am expliciting that we 'spawn' for
all processes. One downside to this is that the startup time is
slower (because we are restarting the process from scratch). To account
for this, I add a short sleep in the associated method (monitor.py::_startup_component()).