fix: extend process 'spawn' delay
In the prior CL, we added a spawn delay of 0.01s, which worked great on
a Linux laptop plugged in. When tested on a Windows laptop unplugged, it
required closer to 1s!

To be fair, it's a slow laptop... To be safe, I am just making the delay
a constant 1s, and adding an explanation. If we find this is very slow,
we can always make this a parameter.