temp fix: remove annoying deprecation warning
We use xarray, which uses pandas, which introduces a DeprecationWarning
to inform users that in an upcoming release, it will require PyArrow.
This is all fine and dandy, but *really* not a concern of ours. Ignoring
DeprecationWarnings because of this, *for now*.

TODO: Revert this when the deprecation warning is removed!