feat: update log handling for 'spawn' process
Since we switched to using the 'spawn' mode of creating new processes,
the logs of child processes were lost. This is because the logger was
not necessarily initialized in the child processes!

We update this by pulling _set_up_logging() from spawn into log, and
passing it as an argument to parser.construct_and_run_component(). This
way, the AfspmComponentsMonitor feeds the log parameters to the
processes as they are constructed.

Additionally, we made changes to our samples since LOGGER_ROOT is now in log.py.

** WARNING **
I should note that is *not* necessarily proper! Logger explicitly warns
against having loggers from different processes printing to a single
file, as it does not inherently support it. We have not seen this issue
during testing, but we *may* need to provide an update if we do.
** END WARNING **