feat: support multiprocess logging with pytest
when 'spawning' multiple processes, their logging setup is not shared
with the new child process (as they are not forked, so the logging
variables are not copied). We had solved this for afspm, but not for
pytest. Pytest has its own logging handler that it spawns, which is not
readily accessible via their API.

The workaround is to allow us to implement *our* logging, according to
the cli-log-level fed to pytest. We do this. To see our logs, you must
feed the -rP parameters. Note that you also need to use one of the 2
methods in this new module, tests/log.py.