feat: add AfspmLogger for using on multiple PCs
The logging functionality as-is works (although sources online say it
shouldnt!), by having all child processes set up their loggers in the
same way. The risk of having multiple processes send logs to a single
file are race conditions, and logging says it is not supported.

For now, we will leave it as-is. However, we have implemented an
AfspmLogger, which in principle allows us to receive logs from multiple
processes/PCs and log them without risk of race conditions. For now,
this should be used if we are running an experiment across multiple
PCs (without it, we would have to merge multiple log files).

In the future, if we find that our existing local logger is broken, we
will modify spawn.py to force-use AfspmLogger always.