fix: add better log for import errors with parser
While debugging a new branch, I was getting an unusual import error tied
to a particular component: it was saying the module could not be found.
Of note: no changes had happened in the associated code for months, so
an introduced bug seemed unlikely.

The culprit turned out to be simple: a ModuleNotFoundError was being
triggered *on top* of another ModuleNotFoundError. The module we were
trying to import was failing because an internal 'import' call was
failing. Explicitly: 'import afspm.components.device.controllers.gxsm.GxsmController'
contained within it an 'import gxsmread'. The latter import was failing,
but the logger was reporting the top-level module import error.

As a helper, we are logging *all* the ModuleNotFoundErrors in this
method via logger.trace.