feat: add units to REQ_PARAM
Here, we add a units attribute to our ParameterMsg in REQ_PARAM. The
point is to explicit the units we are feeding when setting anything. As
described in design_philosophy_md, we expect the receiver to convert any
data to their internal reference units, as necessary. This puts the onus
on each component, which should guarantee more stable code overall.

In terms of changes:
- design_philosophy.d explains the units responsibilities.
- control.proto has the change to ParameterMsg.
- in device/params.py, we remove all references to units and change
SCAN_SPEED_NM_S to SCAN_SPEED.
- in gxsm/params.py, we modify the associated methods to deal with
units, and change SCAN_SPEED as above.
- in device/controller.py, we remove unused logic (set_param_map),
change how self.param_method_map is called (to feed the units as well),
and update the description of what an expected callable in
param_method_map should look like.
- in test_controller.py, we update the scan-speed logic according to the
above changes.
- in test_components_logic/sample_components we also make modifications accordingly.