feat: add request_timeout_ms to test_controller
The asylum device seems much slower responding to requests, so we are
making this attribute settable more easily via test_controller. We
expect a 'reqeust_timeout_ms' at the top-level of your config in order
to do this for the client in test_controller, and we follow the standard
of setting the client's request timeout to 2x what is provided.