refactor/fix: move AsylumParams into dict
Because we have parameter strings that have the same value but different
meaning, we cannot use a StrEnum to compare them (this merges
duplicates). Instead, we change AsylumParam to be an IntEnum (default),
and add a PARAM_STR_MAP, to map to the associated strings.

The rest of the changes are around this larger change. Effectively, we
need to extract the string every time we want to call a parameter now,
so get_param/set_param (and the _list() equivalents) need to change to.

Note, also, that we changed AsylumParameter to AsylumParam.