feat: set lastscan and saveimage on asylum startup
When starting up the asylum DeviceController, we need to make sure
that (a) we are set to save scans automatically, and (b) we are only
scanning one time per request. This CL updates the logic to do so (we
were previously only doing (a)).

We simplify the logic a bit, by having one _set_save_params() method,
which receives these two attributes and a statement of whether or not to
store the prior values. This is used in __init__() and at deletion.