feat: improve scan polling logic
In the device controller, we only send out new scans if the new ones are
different from the old ones. Previously, we would set 2
booleans (timestamps_different and values_different), and only consider
the latter if the former was false (presuming there were no timestamps).

Or at least, that was the intention! Looking over the code, we are
checking the scan values every time (because it is an independent bool).

We aim to fix that in this CL. Note that it assumes Python bools behave
like C, meaning that if the first boolean is True, we do not check the rest...