feat: change logic of gxsm/params get_set()
We were previously not allowing get_set to be called with no curr_units.
I don't understand why, since the lower-level method (units.convert)
should handle this accordingly, throwing an exception as needed. So,
removing this logic for now.

We'll see if this will turn out to have been a major mistake!