feat: update units.convert() to force explicitness
We ran into serious bugs because we were calling convert() with units
defined but not desired_units. After reflection, I realized we
should *never* allow one of these to be proved: either the user is
converting or *not* converting, but they must be explicit.

This will ensure we do not pass stuff that we *think* is converted, but
is not (as was the issue).