fix: update common timing for Windows
It appears that Windows needs significantly longer default timeouts /
sleeps on startup than Linux. (Or, at least, the Windows device I was
testing requires this).

In this CL we update the default timings and update one of our
tests (which requires hard-coded sleeps) to use sleeps based on the
default timings.