fix: revert common timing to prior
We had changed the common timings to 'pass' on Windows, thinking this would resolve
stuff everywhere. Alas, it causes similar warnings on Linux!

Looking into it, it seems that both of these warnings are due to our use of
ctx.destroy() when closing certain components. This is a very ungraceful way to end
things, but it is a reliable way to make sure components are closed.

What we *should* do is investigate changing ctx.destroy() to ctx.term(); doign so should
stop the previously mentioned warnings. I am adding a TODO in this CL to do so.