refactor: rename Device/AfspmController (docs)
As a first pass to rename all DeviceController and AfspmController
logic, we are updating the docs accordingly right now. For now, we are
thinking of renaming DeviceController to 'MicroscopeCommunicator', and
AfspmController to 'MicroscopeProxy'. The former is the only
communicator with the actual microscope / microscope controller. The
latter is a 'proxy' to ensure only one device accesses the microscope at
a time, and to cache information.