doc: update docs to match paper draft
I did a first pass on a potential paper for this framework. I'm updating
the main presentation and code diagrams to more closely match it.

Note that we changed the way aspm_hl_example looks, to more closely
match our point_subscan/grid_subscan examples. This hopefully makes
demoing it easier.

Also note that in code.org the main changes are to replace plantuml
variables devcon with translator and afspmcon with scheduler.