feat: switch ctrl req to demand exp problem
We used to expect a ControlMode when control was requested. It makes
more sense to demand a 'problem' that your component resolves (with
EP_NONE implying you are a general component). This way, only components
that solve problems that have been flagged can gain control.

This CL switches this logic. In doing so, it affects a ton of
classes/tests that account for this!

Note that we added a method to common.py, to more easily check if a
problem is in the current problems set (which accounts for EP_NONE
meaning no problems in the problem set).