fix: correct control_ui's disabling of modes
The prior logic had control modes disabled based on the mode we are
currently in. In fact, the control mode selection should be based on
whether or not problems have been flagged. If no problems are flagged,
we cannot switch to CM_PROBLEM; if problems *are* flagged, we cannot
switch to CM_AUTOMATED. This CL fixes that.