feat: update heartbeat logic to allow quick close
We have had a consistent issue for a while: calling 'end experiment'
would work, but the monitor would take a long time to actually close.

The source of the issue: in our heartbeat listener, we were grabbing the
top of our socket 'queue' every time, rather than grabbing everything in
the queue. Thus, you could run into situations where a bunch of
heartbeats 'bundled up', causing it to take a long time to get to the
kill signal.

We are fixing this logic by grabbing and flushing the whole queue every
time we poll the heartbeat listener.