[feat] use traceback for logging exceptions
Previously, in the parser package, we were using logging.exception() to
print exception information.  We were *not* logging general exceptions
in component:run().

We are doing two things here:
1. We are changing the exception logging to use the traceback module,
with traceback.format_exc() passed to the logger's error function
instead of logger.exception(). This logging makes sure the full
traceback is colored red, which ... seems better? We may end up
reverting this particular change if we find it has adverse side-effects
we have not yet seen.
2. We are adding generic exception logging to the run() method of
component. This way, we should at least see more info on why a
particular component crashed.