[feat] update pubsub logic to grab all queued msgs
Similar to the changes in 189b7c5, in the pubsub path we were only
grabbing the latest message in our message queues, rather than flushing
the queue every time we polled it.

In this CL, we update to do the latter. We also noticed that we didn't
handle an edge case: when we poll() and receive the KILL signal, we were
returning None and adding that to our list of messages (e.g. we would
end up with a [(str, Message), None, ...]).  This is because
_on_message_received() returns None if the message is a KILL signal. We
updated the logic, so the None is not appended.

Lastly, we updated a comment and logic in test_pubsub_logic, to account
for this new behavior. Specifically, we do not need to poll 2x in one of
the tests anymore.