feat: updating to latest probotuf format style
Here, we are switching to the 'editions' style of protobuffers.  The
main difference in the source is that we remove all 'optional'
indicators, as these are by default.

A potential hiccup is that this requires a newer version of the
protobuffers compiler, protoc.  I used the latest (29.3), which required
me to update my protobuf package to a compatible one (5.29.3).

This may cause issues/annoyances with Python versioning on tested
devices... We will see.