[feat] add graphviz support to visualize configs
Here, we add graphviz as a requirement, and create a new script
graphify.py, to convert TOML config files to a handy visualization.
In this visualization, components are shown along with their IO node
connections to different URLs.
The purpose is to be able to confirm that every component is reading
from/writing to where it should be.

Note that the added package of graphviz does not install a local
graphviz executable! It is the user's responsibility to ensure that
graphviz is installed locally. Instructions for doing so can be found in
the python package graphviz's documentation.