[feat] add angle attribute to ScanParameters2d
It is not uncommon that a user needs to modify the angle of the scan
when running an experiment. Thus, we need to add an angle parameter to
our ScanParams. We have done so in this CL, with some caveats/notes:
- We added support for asylum and gxsm for this, that still need to be
tested! The support involves both setting/getting the angle from the
controller *and* getting the angle from received scans.
- We have modified the common create_scan_params_2d() helper to take in
an angle param, which defaults to 0 instead of None. The idea is
that *at this stage* you should know exactly what angle you are wanting.
- We added a method to warn the user (via the log) if it receives scans
without the angle being set, as this may cause issues in interpretation.