[feat] proper import files in dir of config
Previously, we were unable to spawn an experiment unless we were in the
directory of our config.toml file. Upon investigation, this was due to
the fact that we had local python modules we were trying to spawn in
that directory. Since python was run from a separate directory, we were
not finding the appropriate modules.

This CL fixes that by adding a new method, consider_config_path(). This
method adds the config file's path to the currently loaded PATH
environment variable, allowing local methods/classes to be instantiated.

While looking into this, we also noticed that the _import_from_string()
method was not handling exceptions properly. Particularly, it would
'gracefully' return if an exception occurred. While nice, this meant
that it was hard to debug the source of a component crashing upon
parsing.

To resolve it,  we explicitly pass along final exceptions when we are
unable to import.