[refactor] move test_translator in translators dir
This is done because we are adding new tests in the
tests/components/microscope dir, and the norecursedirs logic of pytest
only allows you to skip directories. Thus, in this CL we move it and
update our pytest.ini file, to ensure we don't run test_translator.py by
default when running pytest.