[feat/refactor] big rewrite translator / params
This CL does *many* things, as a first step to making writing
MicroscopeTranslators easier. First things first: the only
translator that is 'supported' after this CL is our fake
ImageTranslator. Getting the other ones working again will be done in
follow-up refactors.

What we have done:
- Added REQ_ACTION as a way to request actions of the microscope. An astute
reader will note that these 'actions' are equivalent to start/stop_scan() and
REQ_START_SCAN / REQ_STOP_SCAN. Indeed, this is something we will need to update
in subsequent updates.
- Modified MicroscopeTranslator. The base class now has two new methods:
  + on_action_request() -> how do you respond to an action?
  + on_param_request() -> how do you respond to a parameter get/set?
  We also have removed the param_method_map logic. It now resides in a separate
  class, MapTranslator.
- Added a new child class to MicroscopeTranslator, called MapTranslator. It
contains dictionaries param_method_map and action_method_map, which map generic
parameters and actions to different Callables. These are handled by its
implementation of on_param_request / on_action_request, where it looks for the
appropriate Callable from the appropriate dict via the generic key.
- Added a new child class to MicroscopeTranslator, called ConfigTranslator. This
one takes advantage of helper classes ParameterHandler and ActionHandler (to be
described later) to ... handle parameter and action requests! It is called a
ConfigTranslator because both of these helper classes load in TOML configs to
define the generic-to-specific mapping, and any special extras.
- In params.py we:
  + Solidified our generic param enum. It now contains all of the main parameters
  we should be using in our calls.
  + Added various errors/exceptions to clarify issues with parameters.
  + Added the ParameterHandler class, which loads generic-to-specific mappings
  via a TOML config, and helps with extra things: type conversions, range checks
  (to make sure you don't set something outside of a supported range), and units.
- In actions.py (new file), we:
  + Create an enum holding generic action names.
  + Added various errors/exceptions to clarify issues with actions.
  + Added the ActionHandler class, which loads generic-to-specific mappings via
  a TOML config to allow SPM-specific action calling.
- Linked to the above errors/exceptions, we added various ControlResponse values
that map to them in control.proto.
- Added test_params.py and test_actions.py for these.
- In feedback.proto, we modified ZCtrlParameteres to hold the zcrl setpoint and
errorGain. Other classes still need to be updated to handle this.
- For all of the above, we updated ImageTranslator, and test_components_logic /
sample_components for the new logic.