[refactor] remove REQ_START/STOP_SCAN for action
Here, we are removing the REQ_START_SCAN / REQ_STOP_SCAN requests and instead
having them exclusively in the set of actions (i.e. handled by REQ_ACTION).
To accomplish this, we do the following to the various MicroscopeTranslator
base classes:
- MicroscopeTranslator: on_action_request() explicitly maps to on_start_scan()
and on_stop_scan() for those actions, and returns a
REP_ACTION_NOT_SUPPORTED otherwise. (This is no longer an abstract method).
- MapTranslator: on_start_scan() and on_stop_scan() now do nothing. We create
abstract methods _build_param_method_map() and _build_action_method_map(),
which are expected to be called in the constructor. The main constructor will
validate that these required actions are in the action_method_map before the
constructor finishes.
- ConfigTranslator: on_start_scan() and on_stop_scan() now do nothing. In the
constructor, we validate that ActionHandler has actions for our required
actions.

To enable REQ_ACTION, we:
- Add a REQ_ACTION to control_pb2.ControlRequest.
- Add a request_parameter() method to io/control/client.py.
- Add a key:val pair to the REQ_TO_OBJ_MAP in io/control/commands.py (to
know what proto is needed for this request).

We also add some TODOs to different documentation, so we (hopefully) don't
forget to update them once this is all done.

For tests, we update:
- sample_components.py: we are moving the parameter / action methods outside of
the class (as one would expect it to be), implementing the
_build_param_method_map (and action equivalent).
- test_scan_handler.py: replace REQ_START_SCAN with REQ_ACTION. Update a test
to consider the expected objects as well.
- test_control_logic.py: update the commands we are testing to no longer test
START_SCAN / STOP_SCAN, but test ACTION and PARAM. We make some code changes
to better support REQ_PARAM, as it (unusually) returns an obj.