[refactor] remove REP_PARAM_INVALID (or action)
We had originally added these to throw an error if the user requested a
parameter or action that was outside of the defined 'generic' ones. However,
doing so limits how easy it is for users to add new actions/parameters. So
now we have removed them, and replaced the errors with logger warnings.