[refactor/feat] add signal stop/start and probepos
Adding the logic for starting and stoping 1D signals (e.g. spectroscopy) and
moving the probe to points in space in order to do so. So we now have 2 new
actions: START_SIGNAL and STOP_SIGNAL, which trigger some action that we expect
to return a Signal1d. Tied to this we have a poll_signal() method, akin to
poll_scan(); and an explicit SS_SIGNALING ScopeState. On the parameter side, we
now have methods to set the probe position. In a lot of cases, the probe
position will be analogous with the top-left position of a scan. However, we
still want these to be explicitly separate to avoid potential issues. Because
of this, we have returned our renaming of SS_COLLECTING to SS_SCANNING (and,
again, have a new SS_SIGNALING).

Many files in here are just tied to renaming SS_COLLECTING to SS_SCANNING.
Below we indicate the other things we've done:

config_translator.py:
- Updated to implement the new abstract methods (and have a default
on_set_probe_pos).

image/translator.py / map_translator.py:
- Updated to implement the new abstract methods.

pbc_logic.py / cache_logic.py:
- adding support for Signal based envelopes, particularly filtering by Signal
type.

test_config_translator.py:
- modified to test probe position logic.

test_control_logic.py:
- modified to test start/stop signal.

test_components_logic / sample_components:
- modified to test setting probe position.