[refactor/feat] add to test_translator (new stuff)
We have a test_translator script that can be used to validate that a given
translator is supportive enough to run. Here, we are adding tests to check for
the new supports we've put in place. Particularly:
- We test start_signal / cancel_signal logic.
- We test ProbePosition setting.
- We check which parameters from params.MicroscopeParameter are supported.
- We check which actions from actions.MicroscopeAction are supported.

For the latter two, we are only checking and printing out via the log. So you
would need to run with --log-cli-level=INFO to get this information. Also, for
checking actions we have added a new method check_action_support() to
MicroscopeTranslator, and ensured the main child classes support it. This
method only sees if the translator 'claims' to support it -- it does not
request the action be done. So the extent to which this works is somewhat
dependent on the developer playing nice. But it's a starting point.