[feat] allow getter or setter in ParamsHandler
Previously, if you were providing methods for getting/setting, you had
to provide both a get and a set. But there are some cases where only a
getter makes sense. So (for now) we allow either a set or get, and send
a warning if you only have one.

If, during usage, you try to call the one you don't have, we throw a ParameterNotSupporterError.