[refactor] first pass gxsm config_translator
This is the first pass to switch the GxsmTranslator to be a
ConfigTranslator. I have not been able to test things, so I am sure
there will be bug fixes shortly.

Once this is working for the original functionality, I will need to add
start_signal/stop_signal/poll_signal logic.