[feat] unit convert -> '' == None
When dealing with empty attributes with protobuffers, if you access them
directly (without checking if they have been set), they will give you a
default version. In the case of str, this is an empty string.

This was causing issues if you are trying to convert units, because we
were treating '' as different from None. Thus, if you tried to convert
from '' to None, it threw an error (because '' was considered a unit).

To get around this, we are now converting any empty strings to None at
the beginning of units.convert.