[fix] remove data_units from poll_scan_params
In config_translator, we were grabinb the units of scan-params 'data'
parameters and feeding those. But the data 'units' parameter should be
NONE at this stage (i.e., also in your params.toml). The only place
where data units should/could be provided is in poll_scans(), because it
is at this point that we have actual scan2ds.

Thus, we are removing this part of poll_scan_params() and updating
test_config_translator accordingly.