[refactor/feat] simplify ActionHandler
Modified the ActionHandler to now handle the approaches indicated by
each child ActionHandler. We do this by making the actions config file a bit
more involved. For example, it now looks like:

        [start-scan]
        method = 'a.b.startscan'
        type = 'PASS_SELF'
        uuid = 'HELLO'

So each generic action is a dict in the config file, and requires at least the
'method' key to be defined. Our parser goes through and imports the method, and
we use 'type' to determine if we need to pass the ActionHandler as the first
input argument to the method.  One can add more input arguments via key:val
pairs (though I suspect it will mostly just be a str like the 'uuid' field.