[fix] change order of setting zctrl params
This is specifically so it works for gxsm. In reality, it may end up
being smarter to override these methods for gxsm, as it is wonky and the
order matters! But since gxsm is the first one we supported, we are kind
of taking it as the 'gold standard' (possibly a mad idea).