[fix/feat] updates test_translator
The main thing is adding logic to handle the case where we change the
offset of our scan region. In this case, we expect the scope state to
change to SS_MOVING until the scope moves.

This was causing issues with our tests, because we were then trying to
reset the scan_params *before* we stopped moving.