[fix] parent/child subscriptions were failing
We expect to be able to subscribe to 'Scan2d' and receive
'Scan2d_Topo-Xp_50.0'. This is natively supported by zmq. However, it
is *not*  natively supported for caching, as this is something we have
written (following guidance from the zmq guide). It looks like we had
introduced a bug in how we handled caching for new subscriptions. This
CL fixes that, so that you can do the above.

We also update our tests to check for this.