[fix] add check_action_support functionality
You had partially added this previously, but had not hooked in
everything on the io side. Now, you can call REQ_ACTION_SUPPORT to check
if a particular action is *supported* before explicitly requesting it be
performed.

This is also used in our test_translator to be able to list supported actions.