[feat/fix] better SS_INTERRUPTED handling
When testing on real devices, we see that we need to set the internal
self.scope_state to SS_INTERRUPTED as well, when an interruption
happens. This allows us to avoid saving scans that were due to interruptions.

However, some devices take longer than our code to update their internal
state. To avoid this issue, we wait one state poll before sending out
the update after our SS_INTERRUPTED.

This makes the code a bit uglier (a stray _was_interrupted variable). So
I may see if I can figure out a better way around it in the future...