[build] update pyproject (gxsmread, project)
We are updating the pyproject to:
(a) Switch to the PEP standard project format (we were using the prior
poetry format).
(b) Switch to the PyPI version of gxsmread, now that we have published
it.
(c) Remove PySimpleGUI, as it is shutting down :(. I have commented it
out, but will need to update to a different GUI library (and update)
control_ui.py accordingly. Until then, it will not work.