[refactor] Signal / SIGNALING -> Spec/SPEC
There appears to be consensus that 'spectroscopy' is a generic term in
SPM for various collections where one variable is changed while multiple
others are recorded. Thus, in this CL we mainly change stuff from signal
to spec.

There is one other major change: we change the format of
Spec1d (previously Signal1d) based on our integration efforts. Earlier,
we explicited 1 independent and 1 dependent variable for spectroscopies.
It turns out that:
(a) it may not be clear from a spec file alone which variable is the
independent/varied one; and
(b) people may be collecting many dependent variables.

Thus, we change the format of Spec1d to have 3 lists for variable names,
units, and data respectively. The data is tabular data.