[feat] first pass enabling spec support gxsm
In this CL, we are enabling spectroscopy support in gxsm. To do so we:
- Updated gxsm/actions.py, as we find an error with the earlier code.
- Uncommented/updated the actions in gxsm/actions.toml.
- Modified gxsm translator's poll_spec() method. It now checks for the
latest filename and loads a spec if available. We also added local
variables last_spec_fname (to avoid loading the same file multiple
times) and old_spec (to hold the old spec in case of issues).