[fix/wip] Set up for saving specs/loading them
- We were not telling gxsm to save specs when they finished. We
are *still* not doing that, but now at least we have a big TODO.
- When getting latest spec filename, we are currently using gxsm's
chfname() method, which checks for SCAN channels. It fails if you do not
have a scan waiting to save (or have done it at least once). Really, we
need a method/way to ask for the save directory *or* the spec filename.