[fix] spec unit tests on gxsm
In the previous push, we were not able to save spectroscopies. As such, our tests
were not all passing. With this CL, we hope to get them all passing.

We now hook into a mechanism to 'autosave' spectroscopy files. This way, when a
spectroscopy ends it is autosaved. Canceled spectroscopies are not saved (dissimilar
from scans - funky, I know).

To avoid duplicating code, we made the SpecModeAction now just a SpecMode, and have
a helper function to convert a spec mode to an 'execute' request.

Lastly, we are putting in big bold letters that GXSM is not expected to perform
spectroscopies before a scan has been done. The net effect is that you cannot expect
the spectroscopy unit tests to pass when run alone on a session startup.