[refactor] initial pass refactoring asylum
Initial pass at refactoring the asylum translator scripts. We have
updated the logic to use ParameterHandler/ActionHandler. However, we
should note that a good deal of base methods in ConfigTranslator needed
to be overriden to handle special aspects of this translator.

There is still a good deal of redundancy to be removed here:
- SCAN_STATUS/SPEC_STATUS are no longer needed. Perhaps
SAVE_IMAGE/SAVE_FORCE either?
- You have overriden poll_probe_pos/on_probe_pos instead of simply
implementing specific setter/getter calls for the
probe-pos-x/probe-pos-y. This means that now, if a user wanted to
individually set probe-pos-x, it will not work.
    + I think you should move these to params.py, which would allow you
    to remove the overrides.
- You are still missing poll_spec()!