[fix] extra boolean check in component
On Windows, we can apparently cause exceptions if we call component's
self.run_per_loop() when we have already received a 'kill' request. The
issue is tied to ZeroMQ's socket handling on closure. To avoid it, we do
an extra self.stay_alive check before calling that specific method.

To be fair, this approach is more proper. I'm just a bit confused as to
why it's an issue on Windows and not on Linux...