[feat] modifying how params works - more flexible
Now, we can provide 'unit' and 'range' without requiring 'uuid' and
'type'. Basically, for a parameter to be supported, it must *either*
have both uuid and type, or have a getter/setter. But the 'unit' and
'range' attributes are kind of independent.

In fact, config_translator assumes that every parameter it queries will
have it's appropriate 'unit'. So this is somewhat important.