[feat] modify test url pub-sub logic
Here, we are modifying the test urls for pub-sub logic for a very
particular reason: they conflict with the default Asylum zmq-xop url.
This means that if you are running the unit tests on the Asylum *while*
your zmq-xop interface is set up, they will fail.

Changing these simply avoids confusion :).