[feat] stop reset SavingMode/ScanningMode on close
We used to be resetting SavingMode/ScanningMode on closure, but this is
causing issues. For now, just removing __del__().

NOTE: If you end up not using this in the future, you may want to modify
the logic by removing the member variables that hold these too. It only
adds complexity.