[feat/fix] remove 'special' set_param_list logic
Previously, we were doing something 'fancy' with set_param_list: we
were *first* converting all values to the SPM-specific units, and *then*
we were setting them. The logic here was that this would allow us to
catch any conversion issues before we set any of the parameters.

The issue is that this logic used set_param_spm(), i.e. the logic that
uses a standard set method. In other words, it did not support
parameters where you have a unique setter!

I started writing a more complicated version of this, which would
consider both versions. Then I realized this is a lot of added logic for
questionable benefit. So, with the aim of following the KISS principle,
I am simply removing the complicated logic.

Now, set_param_list() just goes through each parameter and explicitly
calls set_param().