[feat/fix] inelegant fix to get method type asylum
Asylum has parameters like 'SaveImage' which exist *twice*: once as a
variable and once as a string. To try and differentiate these, we had
set up a key:val mapping that mapped a uuid to a type. The problem is,
we put the SPM uuid as the type (i.e. SaveImage)! So we were just
overwriting it and, well, not solving our problem.

To fix this, we change the key:val mapping to hold the *generic* uuid as
the key. The next problem is that get_param_spm()/set_param_spm() (the
spm-specific methods) only receive the SPM uuid. To work around this, we
override the higher-level get_param()/set_param() methods. Rather than
copy-pasting and modifying the whole code, we simply...store the get_set
method in an ugly member variable before calling the parent
get_param()/set_param().

So this solves our problem without duplidating code. But I certainly
would not call it elegant. So it goes.